Simplified the comparison function when finding annotaions that overlap.
Previously, we were finding overlaps by searching up/downstream of an
initial overlapping feature until no more could be found. However, the
ordering of these features can get messed up when encountering genes
within other genes and can create problems when searching by index in
a list of features. To circumvent this, we now search for all overlaps
throughout the entire list of features.

Additionally, the order in which features are categorized matters when
considering annotations with multiple conflicts. Previously we were
missing the counts of some annotations if there were multiple
conflicting overlaps. We now successfully do the following in order:
1) Check all overlaps for an exact match. If one is found move on to next
   feature.
2) If no exact match can be found, report all in frame overlaps as
   conflicts.
3) If no conflicts can be found, report as a unique annotation.