Implement IntervalTree data structure for overlaps
The IntervalTree package drastically simplifies and
speeds up the process of finding overlaps, exact matches,
and unique genes. This replaces the need to loop through every
feature over and over...Also changed the name of the evidence code
attribute attached to each feature and modified the names of the
returned lists in the compare function for consistency.