compare: Add a column to the colo and confl tsv files for easy grepping
It's useful to point out which conflicting and co-located pairs
have the differentially named genes. While it's not too hard to write an
awk one-liner to get them, it's something many people are potentially interested
in, so I'd like to make it simpler to pull out. I wanted to avoid creating
another set of files that would just be a subset of the existing reports,
so here I instead create a new column with easily-greppable tag strings that
label these properties. I also label the case of differential pseudogene calls
since these are also potentially interesting.