Include pseudogene report summary in write_reports()
Removes the need to call write_reports twice. We are only
including a summary report for pseudogenes so we no longer
need the 'suffix' parameter. Explicitly define what will be printed
to the summary report so future changes are easier to trace. We now
make the distinction between "conflicting_genes" and "conflicts" where
"conflicting_genes" represents a unique set of conflicting genes and
"conflicts" represents the total amount of conflicts. To facilitate this
distinction, the get_named function now has an additional parameter,
"unique", which will return a unique set of features.