Update coordinates in valid transl_except qualifiers
RATT will carry over the transl_except qualifier from the
reference annotation, but the coordinates are representative of
the translation exception in the reference genome. We only care
about translational exceptions that recode stop codons
(i.e. selenocysteine amino acid).This will update
those coordinates to the corresponding location in the query genome
if there's actually a stop codon in that position. Otherwise,
leave it alone and it will be removed later on in pseudoscan if warranted.