HybGenBankWriter: use deepcopy() instead of copy() for temp_local
When we try to temporarily remove the .ref attribute on feature locations
prior to writing genbank output, we used a regular copy() instead of
a deepcopy(). For CompoundLocations, this is insufficient because it's
composed of multiple SimpleLocations. A regular copy maintains references
to the original component locations and they're irreversibly changed.