stopseeker: use a proper gene as test case for minus_strand; fix off-by-one
I introduced an off-by-one error to correct for the minus_strand test case,
but this was because the gene I used turned out to be a pseudo (not divisible
by three). I caught the problem while trying to reconcile with the pseudoscan unit test that began failing when I introduced the error.