has_broken_stop() should check for transl_except attribute
Instead of relying on the 'transl_except' qualifiers in a feature,
we should be using the transl_except attribute. This helps distinguish
between features with valid and invalid selenocystein proteins without
having to rely on the feature qualifiers. Also will prevent false
positive 'fake stops'. ex) Although very unlikely, the reference
transl_except qualifier location could overlap in frame by chance with a feature.
But we won't have to worry about that anymore because the transl_except
attribute is verified by the alignment in coord_check.