Update translation qualifier for valid transl_except features
We check to make sure that valid transl_except amino acids exist
in the sequence, but we never saved that modified sequence.
Now we update the translation qualifier to preserve that information.
This will work regardless if the feature has been unmodified,
corrected, or reverted because the final feature (and it's qualifiers)
is determined by pseudoscan (ie og_feature vs feature).