Update tranl_except translations to mimic to_stop=True
Reorganized how we handle tranl exceptions in has_broken_stop. Now,
everything is under one if statement.

For all other annotations, the translation is generated with to_stop=True
- meaning the final stop codon is ommitted in the string of the translation qualifier.
When updating the translation qualifier of a tranl_except gene we now
remove the final stop codon to be consistent with all other annotations.
This is necessary because this is the only time a translation is directly modified
in the pipeline.