Mistakenly modified 'translation' variable
The translation variable is used at the end of has_broken_stop to
determine the number of internal stops. By changing this variable in
the "transl_except" section, the updated translation was not being considered
at the end. Now we properly keep track of the updated translation string, and
we update the translation qualifier with this string minus the last
stop codon.