test_BLAST: Correct input to reference_match() and update expected results
Expected results manually verified for rplB. The mistake was in
passing a SeqRecord of the SeqFeature object as input to
NcbiBlastpCommandline. This wasn't supposed to work, but a mangled
string resulted from it that was somehow parsed by Biopython/BLAST as
a seemingly longer sequence, but it was not a real or valid one. The
result was that query coverages were read as lower than they should
have been and even the case of rplB, which was completely identical.

This input error only occurred in the unit test. All invocations of
reference_match in the main code base uses correct inputs.