Override Bio.Seq.translate() to avoid `cds` and `to_stop` conflicts
We almost always want the behavior of cds=True because it handles
alternative methionine codons when they're at the start of the
sequence. But this option causes an exception to be thrown if there
are internal stop codons in the sequence, but we want to be able to
translate those, too, in the same way.