WIP: compare(): distinguish between nonconflicting and nonoverlapping features
For the purposes of `hybran compare`, we consider both of these
categories to be "unique". For the purposes of annomerge, we
investigate all cases of overlaps, so we now separate them in the
output of the compare() function. Hybran compare's main() combines
them back again to maintain the intended behavior of that command.

Remaining to do:
- Fix test_compare() for new behavior.