compare: add option 'eliminate_colocated' and enable it for annomerge
If a pair of genes are colocated, we considered them both to be
accounted for, so did not explore further conflicts. For annomerge,
this is problematic since, if a single annotation itself contains
features conflicting with each other, we don't have them flagged for
resolution when they also conflict with a colocated feature.