Properly reset pseudo status and when notes should be written for fusions
After the fusion_upgrade() call, we had leftover evidence codes from the
CDS that only applied to it with its old name. Furthermore, we had the
pseudoscan barcode included in the final annotation when the feature was
uninteresting (all metrics passing). This was due to missing blast
information being considered interesting.

co-authored with Theron James