Prevent empty translations from crashing Hybran
Indexing the last element in a translated string can
cause an error if that string is empty ie) 'Rv3922c' from the
nanopore assembly of FAX57845. In this instance, RATT makes a
bad call and we correct the gene's stop position in a way that
generates a feature with only 1 base pair - which can't translate to
anything. We account for situations where a correction causes the
end to be earlier than the start (and vice versa), but haven't
encountered (until now) a situation that results in a feature with an EMPTY
translation. This should prevent that error from happening.