coord_check: allow for fix_stop to correct genes with delayed ends
When setting seek_stop to False, we indicate that we're not interested
in an intact coding sequence. For when we want to know a gene's
reference-corresponding stop when it has a delayed end, we can make
use of relaxed_found_high (which doesn't require a proper stop codon)
and seek_stop==False to find it. This is useful for whole-gene fusions.