Fix issue with unclustered (due to no blast hits) genes not being named
An example of this is L_00301 in 1-0006 (loc 334956-335846, -1). It
has no blast hits to even itself under our cautious blast
settings (`-seg yes -soft_masking true -comp_based_stats 2`) despite
not screening the hits for any thresholds.

So without appearing in the all vs all blast results, MCL is
completely unaware of it, so we need to add in to the results as
singleton genes those from the all vs. all blast inputs that we don't
see in the clustering results.