Change default blast thresholds back to 95%
My rationale for lowering them to 90% was that, since fixing the
CD-HIT application to remove redundancy only, MCL has more of a role
in making calls. However, it's looking more and more like MCL isn't
suitable for clustering at the granularity that we care about (even with
higher inflation value settings), so there's not really much of a
reason to do so.